+++ src/datasource.ts
@@ -11,6 +11,8 @@ import {
 } from '@grafana/data';
 import { getBackendSrv } from '@grafana/runtime';
 import { filter, isEmpty } from 'lodash';
+// eslint-disable-next-line  no-restricted-imports
+import moment from 'moment';
 import { catchError, lastValueFrom, map, Observable, of } from 'rxjs';
 
 import {
@@ -88,6 +90,7 @@ export class DataSource extends DataSourceApi<
   async query(options: DataQueryRequest<BitmovinAnalyticsDataQuery>): Promise<DataQueryResponse> {
     const { range } = options;
     const isRelativeRangeFrom = this.isRelativeRangeFrom(range.raw);
+    let momentTimeUnit = undefined;
 
     //filter disabled queries
     const enabledQueries = (options.targets = filter(options.targets, (t) => !t.hide));
@@ -101,7 +104,9 @@ export class DataSource extends DataSourceApi<
           ? calculateQueryInterval(target.interval, range!.from.valueOf(), range!.to.valueOf())
           : undefined;
 
-      let queryFrom = range!.from;
+      // create new moment object to not mutate the original grafana object with startOf() to not change
+      // the grafana graph at this point as this would change the timeframe for all the following queries
+      let queryFrom = moment(range!.from.valueOf());
       const queryTo = range!.to;
 
       // floor the query start time to improve cache hitting
@@ -111,11 +116,9 @@ export class DataSource extends DataSourceApi<
           // to allow higher granularity if interval is selected by user
           flooringInterval = getSmallerInterval(interval, flooringInterval);
         }
-        const momentTimeUnit = getMomentTimeUnitForQueryInterval(flooringInterval);
+        momentTimeUnit = getMomentTimeUnitForQueryInterval(flooringInterval);
         if (momentTimeUnit != null) {
-          // range from is a moment and startOf is mutating moment object so this has a side effect to also change the
-          // grafana selected timeframe value which will adapt the grafana graph as well
-          queryFrom = range!.from.startOf(momentTimeUnit);
+          queryFrom.startOf(momentTimeUnit);
         }
       }
 
@@ -202,6 +205,11 @@ export class DataSource extends DataSourceApi<
         meta: { notices: metaNotices },
       });
     });
+
+    // round down grafana start time to adjust the grafana graph and show first data point
+    if (momentTimeUnit != null) {
+      range.from.startOf(momentTimeUnit);
+    }
 
     return Promise.all(promises).then((data) => ({ data }));
   }